shell.req: correctly detect #!/bin/env bash (ALT#35376)
The patterns used were OK for the output of "file -NF$'\t'": it would
put a space after the separator (the default separator being ":").
But not for the output of "file -b": we need to pad the result with a
space to use the same patterns.

Putting a space at the beggining is convenient, because it allows to
match independently for "* sh" or "* bash", otherwise "*sh" would
consume "bash", too.

Other uses of "file -b" in scripts/ don't suffer from this problem.