macro.c: force GNU getopt(3) to reinitialize it's internal data structure
getopt(3) holds pointers to previous argument vector in case you want to
continue or restart parsing.
In grabArgs function argument vector is on the stack, so you are lucky
if this memory is not reused for something else between subsequent calls
to the function.  If you are not lucky, getopt(3) will read new
arbitrary value from the stack and return error.