rpmds.c: rpmRangesOverlap() made asymmetric wrt underspecified release, disttag
To be overlapping, an underspecified release or disttag is admissible
only in Requires, but on the side of Provides. (There was a similar behavior
w.r.t. epoch before.)

Overlapping example (worked before and works now):

Provides: N = V-R
Requires: N = V

Non-overlapping example (used to overlap, i.e. to satisfy the Requires):

Provides: N = V
Requires: N = V-R

We don't want this Requires to be satisfied in this case.