rpmds.c: rpmdsCompareEVR() simplified by moving vars into a more local scope
The moved variables are used only in the inner scope. This way the
code is a bit easier to read and understand.

Also, these pointers are the only pointers there to access the results
of parseEVRD(), so I've marked them the "restrict" keyword. However,
this doesn't lead to noticeable optimization with gcc on x86_64.