build/interdep.c: always fix interpackage deps that need Epoch or Disttag
This completes the intention of de77ec455683b9c1d650c9e1f60daaa6567951f3
"build/interdep.c: fix interpackage dependencies automatically":

> ...This change also downgrades "dependency needs Epoch" error to warning
> because it automatically adds strict dependencies.

However, the dependency pkg1->pkg2 was not fixed (if it needed an epoch
or disttag) if pkg1 was already considered to have a strict transitive
dependency on pkg2 (perhaps, via another pkg3: if pkg1->pkg3 and
pkg3->pkg2 deps have already been fixed and added as strict ones).

Then, if pkg3 was a part of a cycle, the pkg1->pkg2 dep wasn't
optimized out. So, it was left untouched, not in the nice format of a
strict dep (i.e., missing an epoch or a disttag).

This was not dangerous, because the strictness was guaranteed by the
transitive dependency, but it was not nice, and was inconvenient for
tools which happen to compare the dep versions as whole strings (say,
for rpm and apt optimizations, or for disttag-unaware old rpm or apt).