rpmEVRDTCompare() added; rpmVersionCompare() rewritten (equivalently) with it
rpmEVRDTCompare() is to be used by APT (which doesn't store the information
internally as RPM headers).

I believe that it's more convenient to have the integer input args as
a basic generic integer type; the comparison algorithm doesn't rely
on a specific size of the integer arguments. I chose the type to be
large enough to store 64 bits (may be a reasonable representation of
time) and to hold the output of strtoull() (a convenient function for
parsing).

And rpm-4.13 uses uint64_t headerGetNumber() a lot to get such values,
i.e. can only give unsigned 64bit values.

I believe there can be no packages with negative buildtime or epoch,
so I make these arguments unsigned in this version, too.

They are optional, and so they are accompanied by booleans:
has_epoch, has_buildtime.

(Previously, I attempted to use the ugly representation as a pointer
and represent the absent value as NULL. Another approach would be to
use a signed integer and use -1 as an indication of an absent value.
But that might not work in future if we need to apply this algorithm
to 64bit unsigned time value. (Who knows... Now the time in RPM is
32bit.))

The added headerGetEVRDT() (used in rpmVersionCompare()) is based on
commit 76debece95c80c0da8ab90a202f1a2093c7eab41 of rpm-build.