rpmEVRDTCompare() & rpmVersionCompare(): fixed the empty vs 0 epoch distinction
The previous treatment of an absent Epoch (as the least one)
lead to 0:1-alt1 being "newer" than 2-alt1.
(Could be checked with rpmevrcmp tool.)

rpmevrcmp from rpm-4.13.0.1-alt6 gives the opposite result for
this example (-1), which is more reasonable.

And now we changed this function to treat an empty epoch as 0, too.

(Probably, this change doesn't affect the behavior of rpm-4.13,
because it seems to substitute 0 in such cases before it gets to this
function, but it would affect APT's comparison, because APT is to use
this function.)