Correctly handle permissions for hardlinked files
... even if some of the files are skipped.

We apply permissions only when we actually wrote the content
to the file. So, if the file containing the content is skipped,
we should not just write the content into a used file,
but also set the metadata for it.

This amends 7ab9b19bd74b895450c5247a8da0aa3c68616105.

Fixes https://bugzilla.altlinux.org/32997

Signed-off-by: Ivan A. Melnikov <iv@altlinux.org>