lib/rpmdb.c: handle BuildTime by dbiFindMatches() and dbiFindByLabelArch()
imz@: long long is at least 64-bit on all platforms and can hold what
we need: all uint32_t values or alternatively -1. So we use itto hold
the intermediate value.

long long is a basic type and has a corresponding strtoll() function.
(strtol() giving long is not good for us, because long is usually
32-bit on 32-bit platforms, and signed, so it wouldn't be able to hold
all uint32_t values. It's better to consider time as unsigned 32-bit
value, if 32-bit; this would allow to delay the Year2038 problem.)