lib/rpmdb.c splitEpoch(): use strtoll which can read the whole expected range
strtol() used to read a long, and long is usually 32-bit on 32-bit
platforms (x86). But the range that we want to read is up to
UINT32_MAX, so a signed long cannot hold all the range.

Whether we can read the whole expected range used to depend on the
platform.

By using strtoll(), we can read the whole expected range of values on
all platforms. (long long is at least 64-bit.)

(I believe that to make things simpler to understand, we should use
the same type to hold the read integer value in this file everywhere.
I'm going to switch from int64_t to long long in the next commit.)