lib/rpmdb.c dbiFindMatches(): make the types to pass the input value match its source
Since we read epoch with strtoll(), it's known to be long long.

long long is a basic type and is at least 64-bit. (And splitEpoch()
checks that it's in the range from 0 to UINT32_MAX.)

It's cleaner if we pass it in a variable whose type is called long
long everywhere (instead of int64_t), and then compare it to whatever
integer value we get from a query (uint64_t from rpmtdGetNumber();
note that it originates from a 32-bit value actually).

We don't use an unsigned type here because the convention is to pass
an absent value as -1. If the convention was different, we could use
unsigned long long (or even unsigned long, because it's good to
holding values up to UINT32_MAX.)