Provides: RPMQ(EPOCH)
To be required by APT (and other tools such as update-kernel) that
need to query the RPM db by specifying N-EPOCH:V-R.

That's a feature that has always been present in rpm-4.13, but not in
all releases of rpm-4.0.4. So, when copying packages (such as
update-kernel) from Sisyphus to p8, a dependency on this feature will
ensure that the tool works correctly.

# Conflicts:
#	alt/rpm.spec